+++ include/Vector.h
@@ -11,19 +11,19 @@ class Vector: public Tuple
 public:
   Vector()=default;
   Vector(const Tuple& a) : Tuple(a) {}
-  Vector(float x, float y, float z): Tuple(x, y, z, 0.0)
+  Vector(double x, double y, double z): Tuple(x, y, z, 0.0)
   {}
-  float magnitude() const
+  double magnitude() const
   {
     return sqrt(pow(x, 2) + pow(y, 2) + pow(z, 2));
   }
   Vector normalize() const
   {
-    float mag = this->magnitude();
+    double mag = this->magnitude();
     Vector v(this->x/mag, this->y/mag,this->z/mag);
     return v;
   }
-  float dot(const Tuple& rhs) const
+  double dot(const Tuple& rhs) const
   {
     return this->x*rhs.x + this->y*rhs.y + this->z*rhs.z + this->w*rhs.w;
   }
